+++ recipes/default.rb
@@ -18,17 +18,6 @@
 # limitations under the License.
 #
 
-package 'unzip'
-package 'libtool'
-package 'rsync'
-package 'autoconf'
-package 'make'
-package 'gcc'
-package 'autogen' unless platform_family?('rhel', 'fedora')
-package 'gtar' if platform?('freebsd')
-package 'cmake'
-#package 'cmake-curses-gui' 
-
 Array(node['ark']['package_dependencies']).each do |pkg|
   package pkg
 end